Attempt to fix iOS 8 layout issue with prefixing
This adds postcss to metalsmith to autoprefix CSS. However autoprefixing
for iOS 8 fixes the issue on the Trello card but causes the homepage header
to break. Solution would be to rewrite and test the Sass against iOS8 (but
not a priority right now).

Trello: https://trello.com/c/wQpEswvv/559-check-whether-ds-pane-layout-is-issue-with-ipad-or-the-layout