Set up search.js to fetch index and run autocomplete
Using XmlHttpRequest to load the JSON search index, Accessible
autocomplete to handle the search input and LunrJS to seach for the
queries in the index and return matching results.

Based on the status of XmlHttpRequest we show a different "No results"
message if user starts typing before the search index has downloaded.

When the index has fully downloaded we re-run complete the query and
return results if any.